refactor: from quizService to timerService
